Fix docker image build issues related to node-gyp
Tried a number of fixes to resolve with ubuntu 18.04 without success.

Update to Ubuntu 19.04 (disco).

Install rsync as it is a dependency of npm.

Remove override of nodejs version, it isn't necessary anymore.

Note that npm reports that node v10 isn't compatible but this is a known issue being fixed in ubuntu

Docker image builds successfully now and confirmed that the openwrt build completes successfully as well.